requirements.txt
gradio>=4.31.0
openai>=1.30.0
anthropicbedrock>=0.2.0; python_version>='3.9'
anthropic>=0.25.0
google-generativeai>=0.5.0
groq>=0.9.0
requests>=2.31.0
python-dotenv>=1.0.0
github3.py>=4.0.1
pydantic>=2.6.0
